requirements.txt
openmeteo-requests
requests-cache
retry-requests
numpy
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib